+++ src/pages/draw/components/drawToolbar/components/tools/arrowTool/index.tsx
@@ -1,6 +1,7 @@
 import { Button, Flex, theme } from "antd";
 import React, { useCallback, useContext, useMemo, useState } from "react";
 import { useIntl } from "react-intl";
+import { DrawStatePublisher } from "@/components/drawCore/extra";
 import { ArrowIcon, LineIcon } from "@/components/icons";
 import {
 	AppSettingsActionContext,
@@ -8,7 +9,6 @@ import {
 } from "@/contexts/appSettingsActionContext";
 import { useStateSubscriber } from "@/hooks/useStateSubscriber";
 import { ToolbarPopover } from "@/pages/draw/components/drawToolbar/components/toolbarPopover";
-import { DrawStatePublisher } from "@/pages/fullScreenDraw/components/drawCore/extra";
 import { type AppSettingsData, AppSettingsGroup } from "@/types/appSettings";
 import { DrawToolbarKeyEventKey } from "@/types/components/drawToolbar";
 import { DrawState } from "@/types/draw";
