+++ src/pages/draw/components/drawToolbar/components/tools/scanQrcodeTool.tsx
@@ -9,11 +9,11 @@ import {
 	useState,
 } from "react";
 import { useIntl } from "react-intl";
+import { DrawStatePublisher } from "@/components/drawCore/extra";
 import { AntdContext } from "@/contexts/antdContext";
 import { useHotkeysApp } from "@/hooks/useHotkeysApp";
 import { useStateSubscriber } from "@/hooks/useStateSubscriber";
 import { DrawContext } from "@/pages/draw/types";
-import { DrawStatePublisher } from "@/pages/fullScreenDraw/components/drawCore/extra";
 import { DrawState } from "@/types/draw";
 import { getPlatformValue } from "@/utils/platform";
 import { zIndexs } from "@/utils/zIndex";
