+++ src/pages/draw/components/selectLayer/index.tsx
@@ -19,6 +19,7 @@ import {
 	getWindowElements,
 	initUiElementsCache,
 } from "@/commands";
+import { DrawStatePublisher } from "@/components/drawCore/extra";
 import { AppContext } from "@/contexts/appContext";
 import {
 	AppSettingsActionContext,
@@ -30,7 +31,6 @@ import {
 } from "@/hooks/useCallbackRender";
 import { useStateRef } from "@/hooks/useStateRef";
 import { useStateSubscriber } from "@/hooks/useStateSubscriber";
-import { DrawStatePublisher } from "@/pages/fullScreenDraw/components/drawCore/extra";
 import {
 	type AppSettingsData,
 	AppSettingsGroup,
